backend/collaboration-service/requirements.txt
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
h11==0.14.0
idna==3.10
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wsproto==1.2.0
